+++ packages/ui-svelte/src/components/ui/command/command-item.svelte
@@ -7,8 +7,9 @@
 		ref = $bindable(null),
 		class: className,
 		children,
+		showIndicator = true,
 		...restProps
-	}: CommandPrimitive.ItemProps = $props();
+	}: CommandPrimitive.ItemProps & { showIndicator?: boolean } = $props();
 </script>
 
 <CommandPrimitive.Item
@@ -21,5 +22,7 @@
 	{...restProps}
 >
 	{@render children?.()}
-	<CheckIcon class="cn-command-item-indicator ml-auto opacity-0 group-has-[[data-slot=command-shortcut]]/command-item:hidden group-data-[checked=true]/command-item:opacity-100" />
+	{#if showIndicator}
+		<CheckIcon class="cn-command-item-indicator ml-auto opacity-0 group-has-[[data-slot=command-shortcut]]/command-item:hidden group-data-[checked=true]/command-item:opacity-100" />
+	{/if}
 </CommandPrimitive.Item>