+++ packages/ui-svelte/src/components/ui/command/index.ts
@@ -4,6 +4,7 @@ import Input from "./command-input.svelte";
 import Item from "./command-item.svelte";
 import List from "./command-list.svelte";
 import Root from "./command.svelte";
+export type { CommandRootApi } from "./command.svelte";
 
 export {
 	Root,